chore(ci): skip cascade if [skip cascade] is found in the PR title
When there are multiple PRs in flight, using a var to disable the
cascading PR that eventually runs the end to end tests (for instance
because the change is about examples) is error prone.

Allowing each pull request to indivudually decide when it is relevant
or not is more practical and fine grain.