Update examples/docker-compose/README.md
The container for the "docker-cli" label defined in the compose file does not contain the bash shell in it, and the forgejo runner defaults to that, with apparently no fall-back to sh if bash is not found in the $PATH.  Thus, the demo_docker.yml example will fail to run unless you specifically define the shell to sh.  This corrects the example to specifically use sh instead of bash for the docker-cli example.