fix: upgrade using a URL instead of a version
It is common to install a version from a branch or a fork. Change the
upgrade to be a URL instead of a version number for simplicity.

Fix a bug by which the lxc-helpers and the runner are upgraded but not
the script itself.

Increase the service unit timeout when stopping the runner: it may
take up to three hours.

Adapt the tests accordingly.