chore: change the license to GPLv3-or-later
The Forgejo runner is part of the Forgejo project. The licensing
agreement in the governance repository reads like this:

> Forgejo accepts contributions compatible with the GPLv3-or-later license.
> The license under which Forgejo is distributed will be changed upon the acceptance of such contributions.

The first step is to update the LICENSE file to reflect that
decision. The individual copyright notice of each source file will be
updated when and if relevant. If a change is made that is eligible for
copyright, the author may decide change the copyright notice from MIT
or Apache 2 to GPLv3-or-later.

Refs licensing agreement https://codeberg.org/forgejo/governance/src/commit/6eb522282f606f65c77b83b662b008aaf444dcfb/AGREEMENTS.md#licensing
Refs blog post about the licensing agreement https://forgejo.org/2024-08-gpl/