feat: wait for services to be healthy before starting a job
If a --health-cmd is defined for a container, block until its status
is healthy or unhealthy. The timeout is defined by the server internal
logic based on associated --health-* defined delays. If it blocks
indefinitely, the job timeout will eventually cancel it.

While waiting, the simplest solution would be to sleep 1 second until
the container is healthy or unhealthy. To minimize log verbosity, the
sleep interval is instead set to --health-interval and default to one
second if it is not defined.

This logic does not apply to host containers as they do not support
services. They are assumed to always be healthy.

If --health-cmd is set for the container running a job, the first step
will start to run without waiting for the container to become healthy.
There may be valid use cases for that but they are not the focus of
this implementation.