chore: require Go v1.24
It is not a breaking change for Forgejo because Forgejo v13 already
depends on Go v1.24 and Forgejo runner v9 is not compatible with
Forgejo versions before v13 from an API point of view. It is however
compatible from a protocol point of view and this does not depend on
the Go version being used.

It would be a breaking change for third party software using the
runner as a package but it is not designed or supported for that.