feat: support actions with 'using: node24'
It's worth noting that runner does not actually invoke different
versions of node depending on the `using` tag -- it just defers to the
`node` command in the path regardless of the tagged value.  But this
change allows `node24` to be provided and run node, the same level of
support as node12...node20.