move mutex to Job
Makes more sense since it is the data object being modified, and although the synthetic test was passing in prev commit the actual tests w/ data race detector were still tripping.  This suggests multiple RunContext objects are in-use with the same Job, which the test has been updated to represent.