fix: the run id is meant to be relative to the repository
It is now the same as `run_number` which is relative to the repository.

The value of `run_id` in GitHub Actions is relative to the repository
and is used by upload-artifact@v4 to build the URL to download an
artifact.

The `run_id` sent by Forgejo is the instance wide unique id assigned to
a run. It cannot be used for to access as it is never exposed by an
API or a web endpoint.

Refs forgejo/runner#187
Refs https://code.forgejo.org/forgejo/upload-artifact/src/tag/v4#outputs
Refs https://docs.github.com/en/actions/reference/workflows-and-actions/contexts#github-context