chore: merge ACT and runner unit & integration tests together
When run in parallel they will race to populate the Go cache in a
different way (GOPROXY is not configured the same) and that will lead
to transient errors when one job tries to use the cache that is
populated differently by the other job.

The chain of `needs` is adjusted accordingly and the required checks too.