chore: call the release notes assistant when opening a new pull request
When a new pull request is opened and has no labels, the release notes
assistant will not be called and it will look like the event was
missed.

Most of the time the label will be set when the pull request is open
and that will fire two runs. However, the release notes assistant is
idempotent and that will just make one extra run in the CI that is a
noop.