fix(pre-commit): don’t default `verbose` to on
`verbose` was only set here as a work-around to make sure users saw the
output of `validate` since there was no fail signal from the executable.

As of https://code.forgejo.org/forgejo/runner/pulls/1009 there now is a
signal that the validation failed which pre-commit can pick up on, so we
no longer need to default `verbose` to be on.