feat: run host child processes in groups
This means process groups on POSIX and "job objects" on Windows. The
idea is that if the command gets cancelled (e.g. due to an Actions job
being cancelled), we need to terminate the entire process tree, not
just the root process.

This missing feature was the cause of forgejo/runner#1038. On POSIX
platforms, cancellation worked essentially by luck: programs tend to
terminate on SIGHUP (that being the default action for that signal), so
when the root process was killed, all of its children would receive
SIGHUP due to losing their controlling TTY and terminate. However, a
program *could* of course handle this signal or detach from the TTY, in
which case cancelation would previously fail to terminate them.

This patch has been tested on both Windows and Linux on the Zig
project's CI runners. Jobs were able to start normally, and could be
cancelled immediately, with no orphaned processes.