feat: include auth token when cloning actions from own instance
Currently, when using an instance with `REQUIRE_SIGNIN_VIEW = true`, it
is impossible to self-host repos containing Forgejo Actions as the runner
does not pass any auth token when it attempts to clone the action.

Let's enable this workflow by passing along the default `github.Token`
token when we're trying to clone the action in question from our own
instance. This can happen in two cases:

1. The remote URL is empty (e.g. `actions/checkout@v5`) and the default
   actions URL is set to one's own instance.
2. The remote URL is set to one's own instance (e.g.
   `https://forgejo.my-instance.com/actions/checkout@v5`).

While the `github.Token` scope is limited to the repo where the action
is running, it does authenticate the runner to the server and allow it
to clone actions repos that are "public" on the instance. However, this
is only true since Forgejo v13.0.0 [1]. As such, only set the token if
the server version is reported as >= 13.0.0.

[1]: https://codeberg.org/forgejo/forgejo/pulls/8889

For more discussion on this feature, see:
https://code.forgejo.org/forgejo/forgejo-actions-feature-requests/issues/3